feat: remove indirect from number_of_keywords
