+++ quotaclimat/data_processing/mediatree/update_pg_keywords.py
@@ -8,13 +8,21 @@ from sqlalchemy.orm import Session
 from postgres.schemas.models import Keywords
 from quotaclimat.data_processing.mediatree.detect_keywords import *
 
-def update_keywords(session: Session) -> list:
+def update_keywords(session: Session, batch_size: int = 50000) -> list:
     saved_keywords = get_keywords_columns(session)
-    logging.info(f"Updating {len(saved_keywords)} saved keywords")
-    for keyword_id, plaintext, keywords_with_timestamp, number_of_keywords in saved_keywords:
-        new_number_of_keywords = count_keywords_duration_overlap(keywords_with_timestamp)
-        logging.debug(f"{keyword_id} new value {new_number_of_keywords}")
-        update_number_of_keywords(session, keyword_id, new_number_of_keywords)
+    total_updates = len(saved_keywords)
+    logging.info(f"Updating {total_updates} saved keywords")
+    for i in range(0, total_updates, batch_size):
+        batch_updates = saved_keywords[i:i+batch_size]
+        for keyword_id, plaintext, keywords_with_timestamp, number_of_keywords in batch_updates:
+            logging
+            new_number_of_keywords = count_keywords_duration_overlap_without_indirect(keywords_with_timestamp)
+            logging.debug(f"{keyword_id} new value {new_number_of_keywords}")
+            update_number_of_keywords(session, keyword_id, new_number_of_keywords)
+
+        logging.info(f"bulk update done {i} out of {total_updates}")
+        session.commit()
+
     logging.info("updated all keywords")
 
 
@@ -34,4 +42,3 @@ def update_number_of_keywords(session: Session, keyword_id: int, new_number_of_k
         {Keywords.number_of_keywords: new_number_of_keywords},
         synchronize_session=False
     )
-    session.commit()
